refactor: modularize workflows, instructions, and client handling #8 #9 #10
Refactored `CounterChatAgent` to use `IChatClient` for better abstraction.
Replaced `CounterChatAgentExtensions` with `AgenticExtentions` for client/auth handling.
Externalized instructions via `InstructionExtensions` and markdown files.
Moved workflow logic to `OrderPlacementWorkflowExtentions`.
Enhanced `RunStreamingAsync` with improved resource handling and error management.
Added OpenTelemetry support for `ChatClient`.
Removed redundant methods and unused dependencies.
