+++ src/CounterService/Agents/CounterChatAgent.cs
@@ -1,25 +1,26 @@
 using System.Diagnostics;
 using System.Text.Json;
+using CounterService.Extentions;
+using CounterService.Instructions;
 using CounterService.Workflows;
 using Microsoft.Agents.AI;
 using Microsoft.Agents.AI.Workflows;
 using Microsoft.Extensions.AI;
 using Microsoft.Identity.Web;
 using ModelContextProtocol.Client;
-using OpenAI.Chat;
 
 namespace CounterService.Agents;
 
-public partial class CounterChatAgent(ChatClient chatClient,
+public partial class CounterChatAgent(IChatClient chatClient,
     IConfiguration configuration,
     IHttpClientFactory httpClientFactory,
     IHttpContextAccessor httpContextAccessor,
-    ILogger<CounterChatAgent> logger) : AIAgent, IDisposable
+    ILogger<CounterChatAgent> logger) : AIAgent
 {
     internal const string AgentName = $"A2A.{nameof(CounterChatAgent)}";
     public static readonly ActivitySource ActivitySource = new(AgentName, "1.0.0");
 
-    public ChatClient ChatClient { get; } = chatClient;
+    public IChatClient ChatClient { get; } = chatClient;
     public IHttpClientFactory HttpClientFactory { get; } = httpClientFactory;
     public IHttpContextAccessor HttpContextAccessor { get; } = httpContextAccessor;
     public ILogger<CounterChatAgent> Logger { get; } = logger;
@@ -34,7 +35,7 @@ public partial class CounterChatAgent(ChatClient chatClient,
         return HttpContextAccessor.HttpContext?.RequestServices.GetService<ITokenAcquisition>();
     }
 
-    internal static string SystemInstructionPrompt => CounterChatAgentExtensions.GetInstruction("CounterChatAgentInstruction");
+    internal static string SystemInstructionPrompt => InstructionExtensions.GetInstruction("CounterChatAgentInstruction");
 
     public override AgentThread GetNewThread()
         => new CustomAgentThread();
@@ -42,9 +43,55 @@ public partial class CounterChatAgent(ChatClient chatClient,
     public override AgentThread DeserializeThread(JsonElement serializedThread, JsonSerializerOptions? jsonSerializerOptions = null)
         => new CustomAgentThread(serializedThread, jsonSerializerOptions);
 
-    public override Task<AgentRunResponse> RunAsync(IEnumerable<Microsoft.Extensions.AI.ChatMessage> messages, AgentThread? thread = null, AgentRunOptions? options = null, CancellationToken cancellationToken = default)
+    public override Task<AgentRunResponse> RunAsync(IEnumerable<ChatMessage> messages, AgentThread? thread = null, AgentRunOptions? options = null, CancellationToken cancellationToken = default)
     {
         throw new NotImplementedException();
+    }
+
+    public override async IAsyncEnumerable<AgentRunResponseUpdate> RunStreamingAsync(
+        IEnumerable<ChatMessage> messages,
+        AgentThread? thread = null,
+        AgentRunOptions? options = null,
+        [System.Runtime.CompilerServices.EnumeratorCancellation] CancellationToken cancellationToken = default)
+    {
+        // Obtain tokens and initialize clients BEFORE streaming starts to avoid scope disposal issues
+        var tokenAcquisition = GetTokenAcquisition();
+        if (tokenAcquisition == null && !IgnoreAuth)
+        {
+            throw new InvalidOperationException("TokenAcquisition service is not available in the current context");
+        }
+
+        var disposables = new List<IDisposable>();
+        var (mcpClient, a2aBaristaAIAgent, a2aKitchenAIAgent) = await InitializeClientsAsync(HttpClientFactory, tokenAcquisition, cancellationToken);
+        if (mcpClient is IDisposable d1) disposables.Add(d1);
+        if (a2aBaristaAIAgent is IDisposable d2) disposables.Add(d2);
+        if (a2aKitchenAIAgent is IDisposable d3) disposables.Add(d3);
+
+        ArgumentNullException.ThrowIfNull(mcpClient);
+        ArgumentNullException.ThrowIfNull(a2aBaristaAIAgent);
+        ArgumentNullException.ThrowIfNull(a2aKitchenAIAgent);
+
+        var workflow = await OrderPlacementWorkflowExtentions.BuildWorkflowCoreAsync(ChatClient, mcpClient, a2aBaristaAIAgent, a2aKitchenAIAgent, workflowName: $"{nameof(CounterChatAgent)}-workflow", cancellationToken)
+            ?? throw new InvalidOperationException("Failed to create workflow");
+
+        try
+        {
+            var lastChatMsg = messages.Where(m => m.Role == ChatRole.User).LastOrDefault()!;
+
+            await foreach (AgentRunResponseUpdate update in ExecuteWorkflowAsync(workflow, lastChatMsg.Text!, cancellationToken: cancellationToken))
+            {
+                yield return update;
+            }
+        }
+        finally
+        {
+            if (workflow is IDisposable disposableWorkflow)
+            {
+                disposableWorkflow.Dispose();
+            }
+
+            foreach (var d in disposables) d.Dispose();
+        }
     }
 
     private async IAsyncEnumerable<AgentRunResponseUpdate> ExecuteWorkflowAsync(
@@ -94,77 +141,15 @@ public partial class CounterChatAgent(ChatClient chatClient,
         }
     }
 
-    public override async IAsyncEnumerable<AgentRunResponseUpdate> RunStreamingAsync(
-        IEnumerable<Microsoft.Extensions.AI.ChatMessage> messages,
-        AgentThread? thread = null,
-        AgentRunOptions? options = null,
-        [System.Runtime.CompilerServices.EnumeratorCancellation] CancellationToken cancellationToken = default)
-    {
-        // Obtain tokens and initialize clients BEFORE streaming starts to avoid scope disposal issues
-        var tokenAcquisition = GetTokenAcquisition();
-        if (tokenAcquisition == null && !IgnoreAuth)
-        {
-            throw new InvalidOperationException("TokenAcquisition service is not available in the current context");
-        }
-
-        var (mcpClient, a2aBaristaAIAgent, a2aKitchenAIAgent) = await InitializeClientsAsync(HttpClientFactory, tokenAcquisition, cancellationToken);
-        
-        var workflow = await GetAgenticWorkflow(ChatClient, Configuration, mcpClient, a2aBaristaAIAgent, a2aKitchenAIAgent, Logger, cancellationToken);
-
-        if (workflow == null)
-        {
-            throw new InvalidOperationException("Failed to create workflow");
-        }
-
-        var lastChatMsg = messages.Where(m => m.Role == ChatRole.User).LastOrDefault()!;
-
-        await foreach (AgentRunResponseUpdate update in ExecuteWorkflowAsync(workflow, lastChatMsg.Text!, cancellationToken: cancellationToken))
-        {
-            yield return update;
-        }
-    }
-
-    private async Task<(McpClient?, AIAgent?, AIAgent?)> InitializeClientsAsync(
+    private static async Task<(McpClient?, AIAgent?, AIAgent?)> InitializeClientsAsync(
         IHttpClientFactory httpClientFactory,
         ITokenAcquisition? tokenAcquisition,
         CancellationToken cancellationToken)
     {
         // Delegate to extension methods - they handle both auth and non-auth scenarios
-        var mcpClient = await CounterChatAgentExtensions.GetMcpClientAsync(httpClientFactory, tokenAcquisition, cancellationToken);
-        var (a2aBaristaAIAgent, a2aKitchenAIAgent) = await CounterChatAgentExtensions.ResolveA2AClientsAsync(httpClientFactory, tokenAcquisition, cancellationToken);
+        var mcpClient = await AgenticExtentions.GetMcpClientAsync(httpClientFactory, tokenAcquisition, cancellationToken);
+        var (a2aBaristaAIAgent, a2aKitchenAIAgent) = await AgenticExtentions.ResolveA2AClientsAsync(httpClientFactory, tokenAcquisition, cancellationToken);
         return (mcpClient, a2aBaristaAIAgent, a2aKitchenAIAgent);
-    }
-
-    public async Task<Workflow?> GetAgenticWorkflow(
-        ChatClient chatClient,
-        IConfiguration configuration,
-        McpClient? mcpClient,
-        AIAgent? a2aBaristaAIAgent,
-        AIAgent? a2aKitchenAIAgent,
-        ILogger<CounterChatAgent> logger,
-        CancellationToken cancellationToken)
-    {
-        if (mcpClient == null || a2aBaristaAIAgent == null || a2aKitchenAIAgent == null)
-        {
-            throw new ArgumentNullException("Required clients cannot be null");
-        }
-
-        // Delegate to extension method for workflow building
-        return await CounterChatAgentExtensions.BuildWorkflowCoreAsync(
-            chatClient,
-            mcpClient,
-            a2aBaristaAIAgent,
-            a2aKitchenAIAgent,
-            workflowName: null,
-            cancellationToken);
-    }
-
-    public void Dispose()
-    {
-        if (McpClient != null)
-#pragma warning disable CA1816 // Dispose methods should call SuppressFinalize
-            GC.SuppressFinalize(McpClient);
-#pragma warning restore CA1816 // Dispose methods should call SuppressFinalize
     }
 }
