+++ src/CounterService/Instructions/CounterChatAgentInstruction.md
